+++ instrumentation/opentelemetry-instrumentation-grpc/src/opentelemetry/instrumentation/grpc/_aio_client.py
@@ -114,11 +114,17 @@ class _BaseAioClientInterceptor(OpenTelemetryClientInterceptor):
             code = await call.code()
             details = await call.details()
 
-            call.add_done_callback(
-                _unary_done_callback(
-                    span, code, details, self._call_response_hook
-                )
+            callback = _unary_done_callback(
+                span, code, details, self._call_response_hook
             )
+            try:
+                call.add_done_callback(callback)
+            except NotImplementedError:
+                # Some grpc.aio interceptors (e.g. interceptors that await the
+                # call object) wrap the call in a type that does not implement
+                # add_done_callback.  In that case call the callback immediately
+                # since code and details are already known.
+                callback(call)
 
             return call
         except grpc.aio.AioRpcError as exc:
@@ -138,9 +144,8 @@ class _BaseAioClientInterceptor(OpenTelemetryClientInterceptor):
             span.end()
 
     def tracing_skipped(self, client_call_details):
-        return (
-            not is_instrumentation_enabled()
-            or not self.rpc_matches_filters(client_call_details)
+        return not is_instrumentation_enabled() or not self.rpc_matches_filters(
+            client_call_details
         )
 
     def rpc_matches_filters(self, client_call_details):
@@ -151,9 +156,7 @@ class UnaryUnaryAioClientInterceptor(
     grpc.aio.UnaryUnaryClientInterceptor,
     _BaseAioClientInterceptor,
 ):
-    async def intercept_unary_unary(
-        self, continuation, client_call_details, request
-    ):
+    async def intercept_unary_unary(self, continuation, client_call_details, request):
         if self.tracing_skipped(client_call_details):
             return await continuation(client_call_details, request)
 
@@ -168,18 +171,14 @@ class UnaryUnaryAioClientInterceptor(
             continuation_with_args = functools.partial(
                 continuation, new_details, request
             )
-            return await self._wrap_unary_response(
-                continuation_with_args, span
-            )
+            return await self._wrap_unary_response(continuation_with_args, span)
 
 
 class UnaryStreamAioClientInterceptor(
     grpc.aio.UnaryStreamClientInterceptor,
     _BaseAioClientInterceptor,
 ):
-    async def intercept_unary_stream(
-        self, continuation, client_call_details, request
-    ):
+    async def intercept_unary_stream(self, continuation, client_call_details, request):
         if self.tracing_skipped(client_call_details):
             return await continuation(client_call_details, request)
 
@@ -212,9 +211,7 @@ class StreamUnaryAioClientInterceptor(
             continuation_with_args = functools.partial(
                 continuation, new_details, request_iterator
             )
-            return await self._wrap_unary_response(
-                continuation_with_args, span
-            )
+            return await self._wrap_unary_response(continuation_with_args, span)
 
 
 class StreamStreamAioClientInterceptor(