+++ instrumentation/opentelemetry-instrumentation-grpc/tests/test_aio_client_interceptor.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from unittest import IsolatedAsyncioTestCase
+from unittest.mock import patch
 
 import grpc
 
@@ -48,9 +49,7 @@ from .protobuf import test_server_pb2_grpc  # pylint: disable=no-name-in-module
 class RecordingInterceptor(grpc.aio.UnaryUnaryClientInterceptor):
     recorded_details = None
 
-    async def intercept_unary_unary(
-        self, continuation, client_call_details, request
-    ):
+    async def intercept_unary_unary(self, continuation, client_call_details, request):
         self.recorded_details = client_call_details
         return await continuation(client_call_details, request)
 
@@ -193,9 +192,7 @@ class TestAioClientInterceptor(TestBase, IsolatedAsyncioTestCase):
         self.assertEqual(len(spans), 1)
         span = spans[0]
 
-        self.assertEqual(
-            span.name, "/GRPCTestServer/BidirectionalStreamingMethod"
-        )
+        self.assertEqual(span.name, "/GRPCTestServer/BidirectionalStreamingMethod")
         self.assertIs(span.kind, trace.SpanKind.CLIENT)
 
         # Check version and name in span's instrumentation info
@@ -264,9 +261,7 @@ class TestAioClientInterceptor(TestBase, IsolatedAsyncioTestCase):
 
     async def test_error_stream_stream(self):
         with self.assertRaises(grpc.RpcError):
-            async for _ in bidirectional_streaming_method(
-                self._stub, error=True
-            ):
+            async for _ in bidirectional_streaming_method(self._stub, error=True):
                 pass
 
         spans = self.memory_exporter.get_finished_spans()
@@ -340,3 +335,33 @@ class TestAioClientInterceptor(TestBase, IsolatedAsyncioTestCase):
 
             spans = self.memory_exporter.get_finished_spans()
             self.assertEqual(len(spans), 0)
+
+    async def test_unary_unary_add_done_callback_not_implemented(self):
+        """Span should still be finished when add_done_callback raises NotImplementedError."""
+
+        def _raise_not_implemented(self_call, callback):
+            raise NotImplementedError
+
+        with patch.object(
+            grpc.aio.UnaryUnaryCall,
+            "add_done_callback",
+            _raise_not_implemented,
+        ):
+            response = await simple_method(self._stub)
+            assert response.response_data == "data"
+
+        spans = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(spans), 1)
+        span = spans[0]
+
+        self.assertEqual(span.name, "/GRPCTestServer/SimpleMethod")
+        self.assertIs(span.kind, trace.SpanKind.CLIENT)
+        self.assertSpanHasAttributes(
+            span,
+            {
+                RPC_METHOD: "SimpleMethod",
+                RPC_SERVICE: "GRPCTestServer",
+                RPC_SYSTEM: "grpc",
+                RPC_GRPC_STATUS_CODE: grpc.StatusCode.OK.value[0],
+            },
+        )